renaming index_pattern folder
