+++ x-pack/plugins/apm/server/lib/data_view/create_static_data_view.ts
