+++ x-pack/plugins/apm/server/lib/data_view/get_dynamic_data_view.ts
