+++ x-pack/plugins/apm/server/lib/helpers/setup_request.test.ts
@@ -25,7 +25,7 @@ jest.mock('../settings/apm_indices/get_apm_indices', () => ({
   }),
 }));
 
-jest.mock('../index_pattern/get_dynamic_data_view', () => ({
+jest.mock('../data_view/get_dynamic_data_view', () => ({
   getDynamicDataView: async () => {
     return;
   },
